requirements.txt
boto==2.38.0
boto3==1.2.3
botocore==1.3.23
click==6.2
docutils==0.12
jmespath==0.9.0
python-dateutil==2.4.2
git+https://github.com/thepwagner/python-systemd.git
more-itertools==2.2
pycrypto==2.6.1
six==1.10.0
wheel==0.24.0
